requirements.txt
alembic==1.10.2
altair==4.2.2
aniso8601==9.0.1
antiorm==1.2.1
attrs==22.2.0
blinker==1.5
blueprint==3.4.2
branca==0.6.0
cachelib==0.10.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dash==2.9.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
db==0.1.1
db-sqlite3==0.0.1
decorator==5.1.1
entrypoints==0.4
Fiona==1.9.3
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
folium==0.14.0
fonttools==4.39.2
geographiclib==2.0
geopandas==0.12.2
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==20.1.0
haversine==2.8.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
matplotlib==3.7.1
mdurl==0.1.2
munch==2.5.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
phonenumbers==8.13.9
Pillow==9.4.0
plotly==5.13.1
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.7.0
Pympler==1.0.1
pyparsing==3.0.9
pyproj==3.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rich==13.3.2
scripts==2.0
semver==2.13.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.6
streamlit==1.20.0
stua==0.2
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
